Move TempfileStreamIterator to opengever.base.
