+++ opengever/disposition/browser/ech0160_export.py
@@ -1,3 +1,4 @@
+from opengever.base.stream import TempfileStreamIterator
 from opengever.disposition.ech0160.sippackage import SIPPackage
 from plone import api
 from Products.Five import BrowserView
@@ -5,8 +6,6 @@ from pyxb.utils.domutils import BindingDOMSupport
 from tempfile import TemporaryFile
 from zipfile import ZIP_DEFLATED
 from zipfile import ZipFile
-from zope.interface import implements
-from ZPublisher.Iterators import IStreamIterator
 
 
 class ECH0160ExportView(BrowserView):
@@ -41,24 +40,3 @@ class ECH0160ExportView(BrowserView):
             package.write_to_zipfile(zipfile)
 
         return tmpfile
-
-
-class TempfileStreamIterator(object):
-
-    implements(IStreamIterator)
-
-    def __init__(self, tmpfile, size, chunksize=1 << 16):
-        self.size = size
-        tmpfile.seek(0)
-        self.file = tmpfile
-        self.chunksize = chunksize
-
-    def next(self):
-        data = self.file.read(self.chunksize)
-        if not data:
-            self.file.close()
-            raise StopIteration
-        return data
-
-    def __len__(self):
-        return self.size
